tools/pyproject.toml
[project]
name = "tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "google-api-python-client>=2.194.0",
    "requests>=2.33.1",
]
[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests*"]
